Adding counter of seconds and adding more translation features

+++ models/audio.rb
@@ -12,6 +12,7 @@ class AudioFile
   field :status, type: String, default: "new"
   field :translator, type: String
   field :file, type: String
+  field :duration, type: Time
 
   def prep_json
     {
@@ -20,7 +21,9 @@ class AudioFile
       translation: translation,
       status: status,
       translator: translator,
-      public_file: file.gsub("public", "")
+      public_file: file.gsub("public", ""),
+      duration: duration.gmtime.strftime('%R:%S'),
+      audio_folder: audio_folder.id
     }
   end
 
@@ -39,7 +42,9 @@ class AudioFolder
   OUTFOLDER = "public/audio"
 
   field :name, type: String
+  field :duration, type: Time
   has_many :audio_files
+
 
   def self.import
     Dir.glob( "#{INFOLDER}/*.zip" ).each do |file|
@@ -51,31 +56,47 @@ class AudioFolder
     name = file.gsub("folders/", "").gsub(".zip", "")
     puts "Handling folder #{name}"
     folder = create( name: name )
+    FileUtils.mkdir_p "#{OUTFOLDER}/#{folder.id}"
+    total = 0
     Zip::File.open(file) do |zip_file|
-      zip_file.each do |entry|
+      total = zip_file.inject(0.0) do |seconds, entry|
         # Extract to file/directory/symlink
 
         if entry.to_s.match(/.wav$/)
           outfile = "#{OUTFOLDER}/#{folder.id}/#{entry}"
-          AudioFile.create!( name: entry , audio_folder: folder, file: outfile ) unless AudioFile.where(name: entry, audio_folder: folder).exists?
+
+
+          audio = AudioFile.create!( name: entry , audio_folder: folder, file: outfile) unless AudioFile.where(name: entry, audio_folder: folder).exists?
           puts "Extracting #{entry.name}"
-          FileUtils.mkdir_p "#{OUTFOLDER}/#{folder.id}"
           entry.extract(outfile) unless File.exists?(outfile)
+
+          wave = WaveInfo.new(outfile)
+          audio.update_attributes!( duration: Time.at(wave.duration))
+          seconds + wave.duration
+        else
+          seconds
         end
+
       end
     end
+
+    folder.update_attributes!( duration: Time.at(total))
     File.delete( file )
   end
 
   def prep_json( page= 1)
+    puts page
     {
+      id: id,
       name: name,
       audios: audio_files.count,
       status: audio_files.where( :status.ne => "translated").none? ? "ready" : "on_process",
       reviewed: audio_files.where( status: "reviewed").count,
       news: audio_files.where( status: "new").count,
       translated: audio_files.where( status: "translated").count,
-      audio_files: audio_files.paginate(page:page, per_page: 30).map(&:prep_json)
+      audio_files: audio_files.paginate(page:page, per_page: 30).map(&:prep_json),
+      pages: audio_files.paginate(page:page, per_page: 30).total_pages,
+      duration: duration.strftime('%H:%M:%S')
     }
   end
 
@@ -87,7 +108,8 @@ class AudioFolder
       status: audio_files.where( :status.ne => "translated").none? ? "ready" : "on_process",
       reviewed: audio_files.where( status: "reviewed").count,
       news: audio_files.where( status: "new").count,
-      translated: audio_files.where( status: "translated").count
+      translated: audio_files.where( status: "translated").count,
+      duration: duration.strftime('%H:%M:%S'),
     }
   end
 
@@ -96,7 +118,7 @@ class AudioFolder
   end
 
   before_destroy do
-    FileUtils.remove_dir("#{OUTFOLDER}/#{id}")
+    FileUtils.remove_dir("#{OUTFOLDER}/#{id}") if Dir.exists?("#{OUTFOLDER}/#{id}")
   end
 
 end
